requirements.txt
fastapi
httpx
uvicorn
hazelcast-python-client
